+++ trino-s3-proxy/src/test/java/io/trino/s3/proxy/server/testing/ManagedS3MockContainer.java
@@ -18,8 +18,8 @@ import com.google.inject.BindingAnnotation;
 import com.google.inject.Inject;
 import io.minio.MakeBucketArgs;
 import io.minio.MinioClient;
+import io.trino.s3.proxy.server.credentials.Credential;
 import io.trino.s3.proxy.server.credentials.Credentials;
-import io.trino.s3.proxy.server.credentials.Credentials.Credential;
 import io.trino.s3.proxy.server.testing.TestingConstants.ForTestingCredentials;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
