+++ trino-s3-proxy/src/test/java/io/trino/s3/proxy/server/testing/TestingS3ClientProvider.java
@@ -16,8 +16,8 @@ package io.trino.s3.proxy.server.testing;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 import io.airlift.http.server.testing.TestingHttpServer;
+import io.trino.s3.proxy.server.credentials.Credential;
 import io.trino.s3.proxy.server.credentials.Credentials;
-import io.trino.s3.proxy.server.credentials.Credentials.Credential;
 import io.trino.s3.proxy.server.rest.TrinoS3ProxyResource;
 import io.trino.s3.proxy.server.testing.TestingConstants.ForTestingCredentials;
 import jakarta.ws.rs.core.UriBuilder;
